backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
pandas==2.2.0
openpyxl==3.1.2
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.27
alembic==1.13.1
httpx==0.26.0
aiofiles==23.2.1
celery==5.3.6
websockets==12.0
python-socketio==5.11.1
prometheus-client==0.20.0
sentry-sdk==1.40.6
structlog==24.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.2.0
ruff==0.2.1
mypy==1.8.0
google-generativeai==0.3.2
cachetools==5.3.2
tenacity==8.2.3
python-magic==0.4.27
chardet==5.2.0